+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.2.0 - 2017-04-11
+### Changed
+- Pass `mode` and `credentials` options to Fetch
+- Switched `Promise` for `async` functions
+
 ## 1.1.0 - 2017-03-27
 ### Added
 - Account methods `forgot()` and `urgency()`